+++ views/settings/create.php
@@ -3,13 +3,13 @@
 use yii\helpers\Html;
 
 /* @var $this yii\web\View */
-/* @var $model wdmg\settings\models\Settings */
+/* @var $model wdmg\options\models\Settings */
 
-$this->title = Yii::t('app/modules/settings', 'Create Settings');
-$this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/settings', 'Settings'), 'url' => ['index']];
+$this->title = Yii::t('app/modules/options', 'Create Settings');
+$this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/options', 'Settings'), 'url' => ['index']];
 $this->params['breadcrumbs'][] = $this->title;
 ?>
-<div class="settings-create">
+<div class="options-create">
 
     <h1><?= Html::encode($this->title) ?></h1>