+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 Changelog
 =========
 
+## 1.1.0 (2019-05-11)
+ * Refactoring: rename module and methods
+ 
 ## 1.0.1 (2019-05-11)
  * Added get/set methods
- * Added settings component
+ * Added options component
  
 ## 1.0.0 (2019-05-10)
  * Added composer package